AlbumView: Give the items proper spacing
GridView does not have a concept of spacing between items, and even if
it did, it wouldn't take spacing between the left and right into
account. With this patch, we expose a minSpacing value which enforces a
min spacing between items of that value. However, it also always tries
to occupy all the given space. So, if the spacing can be increased more
to be make it look pretier, it will do so.