Fork PlasmaExtras.Header
It's a simple component which allows one to set how big the text should
be without having to specify an exact point size.

This component now uses QtControls.Label instead of the Plasma Label and
theme. This isn't perfect as we are still depending on the plasma theme
for the font size, but we are getting there.