ImageGrid: Change the sourceSize
Initially, I didn't really want to change the sourceSize as them loading
the actual images might be slower, but without changing this, one cannot
easily use the scroll bar on the side to skip through an album. It is
just too slow.

At some point we might want to move to KIO's thumbnail job instead.